requirements.txt
numpy==1.26.4
scipy==1.12.0
torch==2.2.1
pyyaml==6.0.1